feat: Create a header component maker function
In this commit, I created a component maker function called Header(). what this
function does is to create a new header component composed of div and within
this div are nested one h1 and 2 span elements. upon invokation of this funct-
ion, the returned component is then appended to the document's div of class
.header-container