feat: Create a card Maker function
In this commit, I created a card component maker function called cardMaker(). it takes
a single object as parameter and populates the created div element with the parameter
object properties.
The component created by the cardMaker() function has a parent div of class .card.
nested with the .card div is another div of class . headline and a div of class .author.
within .author div, there's a div of class . img-container and nested within it is an ima
ge tag/element