feat: Loop through axios resolve promise data
In this commmit, I looped throught the resolve promise data which happens to be a
gigantic object. the object has 5 properties and each of these props are arrays of objects.
the properties within response data are: javascript, bootstrap, technology, jquery and node.
since each prop is an array of objects, i looped through each object and invoked the card
maker function passing each object as argument, thereby, creating a card from each item of
the array