REFACTOR: Pass relevant Object info into cardMaker()
In this commit, the cardMaker's Object properties are modified to be identical with
the result from the axios get request response data. fields/obj props returned from
the axios response data are headline, authorPhoto and authorName. these fields are
then used to populated the created card which is to be displayed in the DOM