REFACTOR: Modify the resolve promise from axios get request
In this commmit, the result of the axios.get() method is refactored by using the
Object.values() method available to every object. This makes it possible to access
the values of each of the object property.

Note that the values of the object being return from the get request are all arrays.
These arrays in turn are arrays of objects, and each object within the array is then
used to create a card by calling the cardMaker() function with it